+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 ### Changed
 
+* Raise an error if `nil` is passed for the `:field` option.
 * Calling `rank!` with an invalid column name, by either specifying the `:field` or `:group_by` option, will not issue a warning anymore. This also means the gem does not require an active database connection when loading classes.
 
 ## [0.2.0] - 2024-08-16